+++ robot-core/src/main/java/org/obolibrary/robot/ReasonerHelper.java
@@ -7,6 +7,8 @@ import org.obolibrary.robot.exceptions.IncoherentRBoxException;
 import org.obolibrary.robot.exceptions.IncoherentTBoxException;
 import org.obolibrary.robot.exceptions.InconsistentOntologyException;
 import org.obolibrary.robot.reason.InferredClassAssertionAxiomGeneratorDirectOnly;
+import org.obolibrary.robot.reason.InferredObjectPropertyDomainAxiomGenerator;
+import org.obolibrary.robot.reason.InferredObjectPropertyRangeAxiomGenerator;
 import org.obolibrary.robot.reason.InferredSubClassAxiomGeneratorIncludingIndirect;
 import org.obolibrary.robot.reason.InferredSubObjectPropertyAxiomGeneratorIncludingIndirect;
 import org.semanticweb.owlapi.model.*;
@@ -321,6 +323,10 @@ public class ReasonerHelper {
         return new InferredInverseObjectPropertiesAxiomGenerator();
       case "objectpropertycharacteristic":
         return new InferredObjectPropertyCharacteristicAxiomGenerator();
+      case "objectpropertyrange":
+        return new InferredObjectPropertyRangeAxiomGenerator();
+      case "objectpropertydomain":
+        return new InferredObjectPropertyDomainAxiomGenerator();
       case "subobjectproperty":
         if (direct) {
           return new InferredSubObjectPropertyAxiomGenerator();